evaluation/LLMAsAJudgeSourceCode/requirements.txt
anthropic
python-dotenv
openai
google.generativeai
pandas
tabulate
scipy